Update flyway due to changes in entities Code style fixes && code clean up Update and fix dates in tables Code name refactoring Fix tests Add new flyway entries to keep flyway updated in accordance to latest changes in entities.
