--- src/main/java/com/school/project/AbstractEntity/AbstractBaseEntity.java
@@ -1,14 +1,0 @@
-package com.school.project.AbstractEntity;
-
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-public abstract class AbstractBaseEntity {
-
-    @Id
-    @Column(name = "ID")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
