+++ src/main/java/com/school/project/base/AbstractBaseEntity.java
@@ -1,0 +1,21 @@
+package com.school.project.base;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@MappedSuperclass
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Setter
+public abstract class AbstractBaseEntity {
+
+    @Id
+    @Column(name = "ID")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+}
