+++ src/main/java/com/school/project/base/AbstractEntity.java
@@ -1,14 +1,23 @@
-package com.school.project.AbstractEntity;
+package com.school.project.base;
 
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.Column;
+import javax.persistence.MappedSuperclass;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import java.util.Date;
 
-public abstract class AbstractEntity  extends AbstractBaseEntity{
+@MappedSuperclass
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Setter
+public abstract class AbstractEntity extends AbstractBaseEntity {
 
     @CreationTimestamp
     @Column(name = "CREATED_DATE", nullable = false, updatable = false)
@@ -19,4 +28,5 @@ public abstract class AbstractEntity  extends AbstractBaseEntity{
     @Column(name = "UPDATED_DATE", nullable = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date updatedDate;
+
 }
