+++ src/main/java/com/school/project/model/entities/Group.java
@@ -1,11 +1,12 @@
 package com.school.project.model.entities;
 
-import com.school.project.AbstractEntity.AbstractEntity;
+import com.school.project.base.AbstractEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.util.Date;
 import java.util.List;
 
 @Entity
@@ -18,6 +19,10 @@ public class Group extends AbstractEntity {
     @OneToOne
     @JoinColumn(name = "MODULE_ID", nullable = false)
     private Module module;
+
+    @Column(name = "START_DATE", nullable = false)
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date startDate;
 
     @OneToMany(cascade = CascadeType.ALL, targetEntity = User.class)
     @JoinTable(name = "GROUP_USER", joinColumns = {@JoinColumn(name = "GROUP_ID")},
