+++ src/main/java/com/school/project/model/entities/Subject.java
@@ -1,12 +1,14 @@
 package com.school.project.model.entities;
 
-import com.school.project.AbstractEntity.AbstractBaseEntity;
+import com.school.project.base.AbstractBaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 
 @Entity
 @Table(name = "SUBJECT")
