+++ src/main/java/com/school/project/repository/GroupRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group,Long> {
+public interface GroupRepository extends JpaRepository<Group, Long> {
 
 }
