+++ src/main/java/com/school/project/service/GroupService.java
@@ -11,4 +11,5 @@ public interface GroupService {
     Group getGroupById(Long id);
 
     List<Group> getAllGroups();
+
 }
