+++ src/main/resources/db/migration/V1_0__INITIALIZE.sql
@@ -5,7 +5,7 @@ create table USER
     LAST_NAME    varchar(50) not null,
     BIRTH_DATE   datetime    not null,
     EMAIL        varchar(50) not null,
-    PHONE_NUMBER varchar(50)      not null,
+    PHONE_NUMBER varchar(50) not null,
     CREATED_DATE datetime    not null,
     UPDATED_DATE datetime    not null,
     primary key (ID)
