+++ src/main/resources/db/migration/V1_2__USER_ACCOUNT.sql
@@ -1,4 +1,4 @@
-INSERT into USER_ACCOUNT(ID,USER_ID,REGISTRATION_DATE,ACCOUNT_ROLE) VALUES
+INSERT into USER_ACCOUNT(ID, USER_ID, REGISTRATION_DATE, ACCOUNT_ROLE) VALUES
 (1,1,'2019-05-04','STUDENT'),
 (2,2,'2019-05-04','STUDENT'),
 (3,2,'2019-05-04','TEACHER  '),
