+++ src/main/resources/db/migration/V1_7__USER_ACCOUNT_FIX.sql
@@ -1,6 +1,6 @@
 UPDATE USER_ACCOUNT
-SET USER_ID =3
-WHERE ID=3;
+SET USER_ID = 3
+WHERE ID = 3;
 
 INSERT INTO USER (ID, FIRST_NAME, LAST_NAME, BIRTH_DATE, EMAIL, PHONE_NUMBER, CREATED_DATE, UPDATED_DATE) VALUES
 (5,'Genadii','Rokov','1992-05-04','rokov_@mail.ru', 017612345678,'2001-02-04','2001-02-04'),
