+++ src/main/resources/db/migration/V1_9__USER_DATES_FIX.sql
@@ -1,0 +1,23 @@
+UPDATE `USER`
+  SET BIRTH_DATE = '1999-04-20'
+  WHERE ID = 1;
+
+UPDATE `USER`
+  SET BIRTH_DATE = '1989-02-11'
+  WHERE ID = 2;
+
+UPDATE `USER`
+  SET BIRTH_DATE = '1995-11-03'
+  WHERE ID = 3;
+
+UPDATE `USER`
+  SET BIRTH_DATE = '1997-01-08'
+  WHERE ID = 4;
+
+UPDATE `USER`
+  SET BIRTH_DATE = '1999-10-28'
+  WHERE ID = 7;
+
+UPDATE `USER`
+  SET CREATED_DATE = '2018-01-09', UPDATED_DATE = '2018-01-09'
+  WHERE ID > 0;
